+++ internal/spot/transport.go
@@ -3,9 +3,8 @@ package spot
 import (
 	"fmt"
 	"net/http"
+	"net/url"
 	"strings"
-
-	"github.com/spotinst/spotinst-sdk-go/spotinst"
 )
 
 const (
@@ -15,54 +14,55 @@ const (
 )
 
 type apiTransport struct {
-	base       http.RoundTripper
-	config     *spotinst.Config
-	identifier string
-	cluster    string
+	roundTripper            http.RoundTripper
+	baseURL                 *url.URL
+	credentials             credentials
+	userAgent               string
+	clusterIdentifier       string
+	clusterUniqueIdentifier string
 }
 
 func (a *apiTransport) RoundTrip(req *http.Request) (*http.Response, error) {
-	creds, err := a.config.Credentials.Get()
-	if err != nil {
-		return nil, err
-	}
 
-	req.Header.Set("Authorization", fmt.Sprintf("Bearer %s", creds.Token))
-	req.Header.Add("User-Agent", a.config.UserAgent)
+	req.Header.Set("Authorization", fmt.Sprintf("Bearer %s", a.credentials.Token))
+	req.Header.Add("User-Agent", a.userAgent)
 
 	query := req.URL.Query()
-	query.Set(queryAccountId, creds.Account)
-	query.Set(queryClusterIdentifier, a.cluster)
+	query.Set(queryAccountId, a.credentials.Account)
+	query.Set(queryClusterIdentifier, a.clusterIdentifier)
 
-	if a.identifier != "" {
-		query.Set(queryKubernetesUniqueIdentifier, a.identifier)
+	if a.clusterUniqueIdentifier != "" {
+		query.Set(queryKubernetesUniqueIdentifier, a.clusterUniqueIdentifier)
 	}
 
 	req.URL.RawQuery = query.Encode()
 
 	// Set request base URL.
-	req.URL.Host = a.config.BaseURL.Host
-	req.URL.Scheme = a.config.BaseURL.Scheme
+	req.URL.Host = a.baseURL.Host
+	req.URL.Scheme = a.baseURL.Scheme
 
 	// Set request headers.
-	req.Host = a.config.BaseURL.Host
+	req.Host = a.baseURL.Host
 
 	if !strings.HasPrefix(req.URL.Path, "/") {
 		req.URL.Path = "/" + req.URL.Path
 	}
 
-	return a.base.RoundTrip(req)
+	return a.roundTripper.RoundTrip(req)
 }
 
-func ApiTransport(base http.RoundTripper, config *spotinst.Config, identifier string, cluster string) http.RoundTripper {
-	if base == nil {
-		base = http.DefaultTransport
+func ApiTransport(roundTripper http.RoundTripper, baseURL *url.URL, creds credentials, clusterIdentifier string, clusterUniqueIdentifier string) http.RoundTripper {
+	if roundTripper == nil {
+		roundTripper = http.DefaultTransport
 	}
+	ua := getUserAgent()
 
 	return &apiTransport{
-		base:       base,
-		config:     config,
-		identifier: identifier,
-		cluster:    cluster,
+		roundTripper:            roundTripper,
+		baseURL:                 baseURL,
+		credentials:             creds,
+		userAgent:               ua,
+		clusterIdentifier:       clusterIdentifier,
+		clusterUniqueIdentifier: clusterUniqueIdentifier,
 	}
 }
